Move FAQCard to domain
