+++ feature/tabsflow/moreflow/faq/src/main/java/com/epmedu/animeal/tabs/more/faq/domain/model/FAQCard.kt
@@ -1,0 +1,7 @@
+package com.epmedu.animeal.tabs.more.faq.domain.model
+
+internal data class FAQCard(
+    val question: String,
+    val answer: String,
+    val isSelected: Boolean = false
+)
