+++ feature/tabsflow/moreflow/faq/src/main/java/com/epmedu/animeal/tabs/more/faq/presentation/FAQScreenEvent.kt
@@ -1,6 +1,6 @@
 package com.epmedu.animeal.tabs.more.faq.presentation
 
-import com.epmedu.animeal.tabs.more.faq.presentation.model.FAQCard
+import com.epmedu.animeal.tabs.more.faq.domain.model.FAQCard
 
 internal sealed interface FAQScreenEvent {
     object BackClicked : FAQScreenEvent
