+++ feature/tabsflow/moreflow/faq/src/main/java/com/epmedu/animeal/tabs/more/faq/presentation/FAQScreenUI.kt
@@ -17,7 +17,7 @@ import com.epmedu.animeal.foundation.theme.AnimealTheme
 import com.epmedu.animeal.resources.R
 import com.epmedu.animeal.tabs.more.faq.presentation.FAQScreenEvent.BackClicked
 import com.epmedu.animeal.tabs.more.faq.presentation.FAQScreenEvent.CardClicked
-import com.epmedu.animeal.tabs.more.faq.presentation.model.FAQCard
+import com.epmedu.animeal.tabs.more.faq.domain.model.FAQCard
 import com.epmedu.animeal.tabs.more.faq.presentation.ui.FAQHeader
 import com.epmedu.animeal.tabs.more.faq.presentation.ui.FAQListItem
 import com.epmedu.animeal.tabs.more.faq.presentation.viewmodel.FAQState
@@ -82,7 +82,7 @@ private fun FAQScreenUIPreview() {
                     FAQCard(
                         question = "Question ${index + 1}",
                         answer = generateLoremIpsum(),
-                        isExpanded = index == 2
+                        isSelected = index == 2
                     )
                 }.toImmutableList()
             ),
