--- feature/tabsflow/moreflow/faq/src/main/java/com/epmedu/animeal/tabs/more/faq/presentation/model/FAQCard.kt
@@ -1,7 +1,0 @@
-package com.epmedu.animeal.tabs.more.faq.presentation.model
-
-internal data class FAQCard(
-    val question: String,
-    val answer: String,
-    val isExpanded: Boolean = false
-)
